refactor: extract setup flow
