+++ views/SetupView.tsx
@@ -1,8 +1,22 @@
 import React, { useEffect, useReducer, useState } from 'react';
 import { ArrowLeft, Swords, Users } from 'lucide-react';
 import { Button } from '../components/ui/Button';
-import { GameConfig, Player, InOutRule, MatchMode } from '../types';
+import { Player, GameConfig, InOutRule, MatchMode } from '../types';
 import type { GameType } from '../utils/arenaFlow';
+import { PlayerNameField } from '../components/game-setup/PlayerNameField';
+import {
+  buildSetupConfig,
+  buildSetupPlayers,
+  createInitialSetupState,
+  deriveSetupLaunchState,
+  getGameName,
+  getMatchModeLabel,
+  getRuleDescription,
+  getRuleLabel,
+  getRulesContent,
+  getSetupTitle,
+  setupReducer,
+} from '../src/features/game-setup/setupModel';
 
 interface SetupViewProps {
   gameType?: GameType;
@@ -29,267 +43,15 @@ const labelClass = 'mb-3 block text-[11px] font-black uppercase tracking-[0.28em
 const activeOptionClass = 'bg-gradient-to-r from-orange-600 to-red-600 text-white border-transparent shadow-[0_0_14px_rgba(234,88,12,0.35)]';
 const inactiveOptionClass = 'bg-white/[0.04] border-white/10 text-gray-400 hover:border-orange-500/40 hover:text-white';
 
-type SetupState = {
-  startingScore: number;
-  customScoreStr: string;
-  matchMode: MatchMode;
-  legsToWin: number;
-  setsToWin: number;
-  cricketRounds: NonNullable<GameConfig['cricketRounds']>;
-  isDoubles: boolean;
-  playerNames: string[];
-  team1Names: string[];
-  team2Names: string[];
-  checkOut: InOutRule;
-  checkIn: InOutRule;
-  startingPlayerIndex: number;
-  teamStarterIds: Record<string, string>;
-  customLegsStr: string;
-};
-
-type SetupAction =
-  | { type: 'apply_game_type_defaults'; gameType: GameType }
-  | { type: 'apply_prefilled_names'; gameType: GameType; names: string[] }
-  | { type: 'apply_prefilled_config'; gameType: GameType; config: Partial<GameConfig> }
-  | { type: 'set_player_count'; gameType: GameType; count: number }
-  | { type: 'update_player_name'; index: number; name: string }
-  | { type: 'update_team_name'; team: 1 | 2; index: number; name: string }
-  | { type: 'update_team_starter'; teamId: 'team1' | 'team2'; playerId: string }
-  | { type: 'set_starting_score'; value: number }
-  | { type: 'set_custom_score_str'; value: string }
-  | { type: 'set_match_mode'; value: MatchMode }
-  | { type: 'set_legs_to_win'; value: number }
-  | { type: 'set_sets_to_win'; value: number }
-  | { type: 'set_cricket_rounds'; value: NonNullable<GameConfig['cricketRounds']> }
-  | { type: 'set_is_doubles'; value: boolean }
-  | { type: 'set_check_in'; value: InOutRule }
-  | { type: 'set_check_out'; value: InOutRule }
-  | { type: 'set_starting_player_index'; value: number }
-  | { type: 'set_custom_legs_str'; value: string }
-  | { type: 'normalize_for_game_type'; gameType: GameType };
-
-const DEFAULT_TEAM_STARTERS = { team1: 't1p1', team2: 't2p1' };
-
-const createInitialSetupState = (): SetupState => ({
-  startingScore: 501,
-  customScoreStr: '170',
-  matchMode: 'LEGS',
-  legsToWin: 3,
-  setsToWin: 3,
-  cricketRounds: 20,
-  isDoubles: false,
-  playerNames: ['', ''],
-  team1Names: ['', ''],
-  team2Names: ['', ''],
-  checkOut: 'Double',
-  checkIn: 'Open',
-  startingPlayerIndex: 0,
-  teamStarterIds: DEFAULT_TEAM_STARTERS,
-  customLegsStr: '7',
-});
-
-const normalizeSetupState = (state: SetupState, gameType: GameType): SetupState => {
-  let nextState = state;
-
-  if (!nextState.isDoubles) {
-    const maxIndex = Math.max(0, nextState.playerNames.length - 1);
-    if (nextState.startingPlayerIndex > maxIndex) {
-      nextState = {
-        ...nextState,
-        startingPlayerIndex: maxIndex,
-      };
-    }
-  }
-
-  if (gameType === 'TRIATHLON' && !nextState.isDoubles && nextState.playerNames.length < 2) {
-    nextState = {
-      ...nextState,
-      playerNames: [nextState.playerNames[0] || '', nextState.playerNames[1] || ''],
-    };
-  }
-
-  return nextState;
-};
-
-const setupReducer = (state: SetupState, action: SetupAction): SetupState => {
-  switch (action.type) {
-    case 'apply_game_type_defaults': {
-      if (action.gameType === 'X01_501_BO5') {
-        return normalizeSetupState({
-          ...state,
-          startingScore: 501,
-          customScoreStr: '501',
-          matchMode: 'LEGS',
-          legsToWin: 3,
-          customLegsStr: '3',
-          setsToWin: 1,
-          isDoubles: false,
-          playerNames: ['', ''],
-          checkIn: 'Open',
-          checkOut: 'Double',
-          startingPlayerIndex: 0,
-          teamStarterIds: DEFAULT_TEAM_STARTERS,
-        }, action.gameType);
-      }
-
-      if (action.gameType === 'X01') {
-        return normalizeSetupState({
-          ...state,
-          startingScore: 501,
-          matchMode: 'LEGS',
-          legsToWin: 3,
-          customLegsStr: '3',
-          setsToWin: 3,
-          isDoubles: false,
-          checkIn: 'Open',
-          checkOut: 'Double',
-          startingPlayerIndex: 0,
-          teamStarterIds: DEFAULT_TEAM_STARTERS,
-        }, action.gameType);
-      }
-
-      if (action.gameType === 'CRICKET') {
-        return normalizeSetupState({
-          ...state,
-          matchMode: 'LEGS',
-          legsToWin: 3,
-          customLegsStr: '3',
-          setsToWin: 1,
-          cricketRounds: 20,
-          isDoubles: false,
-        }, action.gameType);
-      }
-
-      if (action.gameType === 'TRIATHLON') {
-        return normalizeSetupState({
-          ...state,
-          matchMode: 'LEGS',
-          legsToWin: 1,
-          customLegsStr: '1',
-          setsToWin: 1,
-          isDoubles: false,
-          playerNames: ['', ''],
-        }, action.gameType);
-      }
-
-      return normalizeSetupState(state, action.gameType);
-    }
-    case 'apply_prefilled_names': {
-      if (action.names.length === 0) return state;
-
-      if (action.gameType === 'X01' && action.names.length === 4) {
-        return normalizeSetupState({
-          ...state,
-          isDoubles: true,
-          playerNames: action.names,
-          team1Names: [action.names[0], action.names[1]],
-          team2Names: [action.names[2], action.names[3]],
-        }, action.gameType);
-      }
-
-      return normalizeSetupState({
-        ...state,
-        isDoubles: false,
-        playerNames: action.gameType === 'TRIATHLON'
-          ? [action.names[0] || '', action.names[1] || '']
-          : action.names,
-        team1Names: action.names.length >= 2 ? [action.names[0], action.names[1]] : state.team1Names,
-        team2Names: action.names.length >= 4 ? [action.names[2], action.names[3]] : state.team2Names,
-      }, action.gameType);
-    }
-    case 'apply_prefilled_config': {
-      const nextState: SetupState = {
-        ...state,
-        startingScore: typeof action.config.startingScore === 'number' ? action.config.startingScore : state.startingScore,
-        customScoreStr: typeof action.config.startingScore === 'number' ? String(action.config.startingScore) : state.customScoreStr,
-        matchMode: action.config.matchMode ?? state.matchMode,
-        legsToWin: typeof action.config.legsToWin === 'number' ? action.config.legsToWin : state.legsToWin,
-        customLegsStr: typeof action.config.legsToWin === 'number' ? String(action.config.legsToWin) : state.customLegsStr,
-        setsToWin: typeof action.config.setsToWin === 'number' ? action.config.setsToWin : state.setsToWin,
-        cricketRounds: action.config.cricketRounds ?? state.cricketRounds,
-        isDoubles: typeof action.config.isDoubles === 'boolean' ? action.config.isDoubles : state.isDoubles,
-        startingPlayerIndex: typeof action.config.initialStartingPlayerIndex === 'number' ? action.config.initialStartingPlayerIndex || 0 : state.startingPlayerIndex,
-        teamStarterIds: action.config.teamStarterIds ?? state.teamStarterIds,
-        checkIn: action.config.checkIn ?? state.checkIn,
-        checkOut: action.config.checkOut ?? state.checkOut,
-      };
-
-      return normalizeSetupState(nextState, action.gameType);
-    }
-    case 'set_player_count': {
-      const minPlayers = (action.gameType === 'CRICKET' || action.gameType === 'TRIATHLON') && !state.isDoubles ? 2 : 1;
-      const maxPlayers =
-        (action.gameType === 'CRICKET' && !state.isDoubles) ? 3 :
-        (action.gameType === 'TRIATHLON' && !state.isDoubles) ? 2 :
-        4;
-      const newCount = Math.max(minPlayers, Math.min(maxPlayers, action.count));
-
-      return normalizeSetupState({
-        ...state,
-        playerNames: newCount > state.playerNames.length
-          ? Array.from({ length: newCount }, (_, index) => state.playerNames[index] || `Joueur ${index + 1}`)
-          : state.playerNames.slice(0, newCount),
-      }, action.gameType);
-    }
-    case 'update_player_name': {
-      const playerNames = [...state.playerNames];
-      playerNames[action.index] = action.name;
-      return { ...state, playerNames };
-    }
-    case 'update_team_name': {
-      if (action.team === 1) {
-        const team1Names = [...state.team1Names];
-        team1Names[action.index] = action.name;
-        return { ...state, team1Names };
-      }
-      const team2Names = [...state.team2Names];
-      team2Names[action.index] = action.name;
-      return { ...state, team2Names };
-    }
-    case 'update_team_starter':
-      return {
-        ...state,
-        teamStarterIds: {
-          ...state.teamStarterIds,
-          [action.teamId]: action.playerId,
-        },
-      };
-    case 'set_starting_score':
-      return { ...state, startingScore: action.value };
-    case 'set_custom_score_str':
-      return { ...state, customScoreStr: action.value };
-    case 'set_match_mode':
-      return { ...state, matchMode: action.value };
-    case 'set_legs_to_win':
-      return { ...state, legsToWin: action.value };
-    case 'set_sets_to_win':
-      return { ...state, setsToWin: action.value };
-    case 'set_cricket_rounds':
-      return { ...state, cricketRounds: action.value };
-    case 'set_is_doubles':
-      return { ...state, isDoubles: action.value };
-    case 'set_check_in':
-      return { ...state, checkIn: action.value };
-    case 'set_check_out':
-      return { ...state, checkOut: action.value };
-    case 'set_starting_player_index':
-      return { ...state, startingPlayerIndex: action.value };
-    case 'set_custom_legs_str':
-      return { ...state, customLegsStr: action.value };
-    case 'normalize_for_game_type':
-      return normalizeSetupState(state, action.gameType);
-    default:
-      return state;
-  }
-};
 
 export const SetupView: React.FC<SetupViewProps> = ({
   onStart,
   onBack,
-  gameType = 'X01',
+  gameType: selectedGameType = 'X01',
   prefilledPlayerNames = [],
   prefilledConfig,
 }) => {
+  const gameType = (selectedGameType ?? 'X01') as GameType;
   const isQuickPreset = gameType === 'X01_501_BO5';
   const [setupState, dispatch] = useReducer(setupReducer, undefined, createInitialSetupState);
   const {
@@ -312,6 +74,14 @@ export const SetupView: React.FC<SetupViewProps> = ({
   const [isRulesOpen, setIsRulesOpen] = useState(false);
   const [isCustomScoreOpen, setIsCustomScoreOpen] = useState(false);
   const [isCustomLegsOpen, setIsCustomLegsOpen] = useState(false);
+  const launchState = deriveSetupLaunchState({
+    gameType,
+    startingScore,
+    customScoreStr,
+    matchMode,
+    legsToWin,
+    customLegsStr,
+  });
 
   useEffect(() => {
     dispatch({ type: 'apply_game_type_defaults', gameType });
@@ -358,32 +128,9 @@ export const SetupView: React.FC<SetupViewProps> = ({
     if (gameType === 'X01' && ((isCustomActive && !isCustomScoreValid) || (matchMode === 'LEGS' && isCustomLegsActive && !isCustomLegsValid))) {
       return;
     }
-
-    const safeStartingScore = startingScore > 0 ? startingScore : 501;
-    let players: Player[] = [];
-
-    if (isQuickPreset) {
-      players = [0, 1].map((index) => ({
-        id: `p${index + 1}`,
-        name: (playerNames[index] || `Joueur ${index + 1}`).trim() || `Joueur ${index + 1}`,
-        teamId: `p${index + 1}`,
-      }));
-    } else if (isDoubles) {
-      const p1 = { id: 't1p1', name: team1Names[0].trim() || 'Joueur 1', teamId: 'team1' };
-      const p2 = { id: 't1p2', name: team1Names[1].trim() || 'Joueur 2', teamId: 'team1' };
-      const p3 = { id: 't2p1', name: team2Names[0].trim() || 'Joueur 3', teamId: 'team2' };
-      const p4 = { id: 't2p2', name: team2Names[1].trim() || 'Joueur 4', teamId: 'team2' };
-      players = [p1, p2, p3, p4];
-    } else {
-      players = playerNames.map((name, i) => ({
-        id: `p${i + 1}`,
-        name: name.trim() || `Joueur ${i + 1}`,
-        teamId: `p${i + 1}`,
-      }));
-    }
-
-    const config: GameConfig = {
-      startingScore: safeStartingScore,
+    const players = buildSetupPlayers({ isQuickPreset, isDoubles, playerNames, team1Names, team2Names });
+    const { config } = buildSetupConfig({
+      startingScore,
       checkIn,
       checkOut,
       matchMode,
@@ -391,49 +138,10 @@ export const SetupView: React.FC<SetupViewProps> = ({
       setsToWin,
       cricketRounds,
       isDoubles,
-      initialStartingPlayerIndex: isDoubles ? 0 : startingPlayerIndex,
-      initialStartingTeamId: undefined,
-      teamStarterIds: isDoubles ? teamStarterIds : undefined,
-    };
-
+      startingPlayerIndex,
+      teamStarterIds,
+    });
     onStart(players, config);
-  };
-
-  const getRuleDescription = (type: 'in' | 'out', rule: InOutRule) => {
-    if (type === 'out') {
-      switch (rule) {
-        case 'Open':
-          return 'Fin sur n importe quel segment.';
-        case 'Double':
-          return 'Fin classique sur un double ou le bull.';
-        case 'Master':
-          return 'Fin sur un double, un triple ou le bull.';
-      }
-    }
-
-    switch (rule) {
-      case 'Open':
-        return 'Le score commence immediatement.';
-      case 'Double':
-        return 'Un double est requis pour ouvrir le score.';
-      case 'Master':
-        return 'Touchez un double ou un triple pour commencer.';
-    }
-  };
-
-  const getRuleLabel = (rule: InOutRule) => {
-    switch (rule) {
-      case 'Open':
-        return 'Open';
-      case 'Double':
-        return 'Double';
-      case 'Master':
-        return 'Master';
-    }
-  };
-
-  const getMatchModeLabel = (mode: MatchMode) => {
-    return mode === 'LEGS' ? 'Manches' : 'Sets';
   };
 
   const presets = [301, 501, 701];
@@ -441,14 +149,13 @@ export const SetupView: React.FC<SetupViewProps> = ({
   const customLegsValue = parseInt(customLegsStr, 10);
   const hasCustomScoreValue = customScoreStr.trim().length > 0;
   const hasCustomLegsValue = customLegsStr.trim().length > 0;
-  const isCustomScoreValid = hasCustomScoreValue && !Number.isNaN(customScoreValue) && customScoreValue >= 2;
-  const isCustomLegsValid = hasCustomLegsValue && !Number.isNaN(customLegsValue) && customLegsValue >= 1;
-  const isPresetSelected = presets.includes(startingScore);
   const presetLegsOptions = [1, 2, 3, 4, 5];
-  const isCustomActive = !isPresetSelected || startingScore === parseInt(customScoreStr || '0', 10);
-  const isCustomLegsActive = matchMode === 'LEGS' && !presetLegsOptions.includes(legsToWin);
-  const isCustomScoreLaunchBlocked = gameType === 'X01' && isCustomActive && !isCustomScoreValid;
-  const isCustomLegsLaunchBlocked = gameType === 'X01' && matchMode === 'LEGS' && isCustomLegsActive && !isCustomLegsValid;
+  const isCustomScoreValid = launchState.isCustomScoreValid;
+  const isCustomLegsValid = launchState.isCustomLegsValid;
+  const isCustomActive = launchState.isCustomActive;
+  const isCustomLegsActive = launchState.isCustomLegsActive;
+  const isCustomScoreLaunchBlocked = launchState.isCustomScoreLaunchBlocked;
+  const isCustomLegsLaunchBlocked = launchState.isCustomLegsLaunchBlocked;
 
   const handleCustomFocus = () => {
     const value = parseInt(customScoreStr, 10);
@@ -490,89 +197,6 @@ export const SetupView: React.FC<SetupViewProps> = ({
     }
   };
 
-  const getTitle = () => {
-    if (gameType === 'CRICKET') return 'Configuration';
-    if (gameType === 'CAPITAL') return 'Configuration';
-    if (gameType === 'TRIATHLON') return 'Configuration';
-    if (gameType === 'X01_501_BO5') return '501 Double Out';
-    return 'Configuration';
-  };
-
-  const getGameName = () => {
-    if (gameType === 'CRICKET') return 'Cricket';
-    if (gameType === 'CAPITAL') return 'Capital';
-    if (gameType === 'TRIATHLON') return 'Le Triathlon';
-    if (gameType === 'X01_501_BO5') return '501, 3 manches gagnantes';
-    return 'X01';
-  };
-
-  const getRulesContent = () => {
-    if (gameType === 'X01' || gameType === 'X01_501_BO5') {
-      return {
-        title: 'Regles Du X01',
-        items: [
-          'Chaque joueur commence avec un score defini, puis retire les points marques a chaque tour.',
-          'Le vainqueur est le premier a atteindre exactement zero.',
-          `Check-in actuel : ${checkIn}. Le score ${checkIn === 'Open' ? 'commence immediatement' : 'ne commence qu apres une ouverture valide'}.`,
-          `Check-out actuel : ${checkOut}. Il faut donc terminer selon cette regle pour valider la victoire.`,
-          'Si un joueur depasse zero ou termine sans respecter le check-out, le tour est annule.',
-        ],
-      };
-    }
-
-    if (gameType === 'CRICKET') {
-      return {
-        title: 'Regles Du Cricket',
-        items: [
-          'Les cibles sont 15, 16, 17, 18, 19, 20 et Bull.',
-          'Il faut fermer chaque nombre avec trois marques.',
-          'Une fois une cible fermee, les marques supplementaires rapportent des points tant que l adversaire ne l a pas fermee.',
-          'Le joueur gagne quand il a tout ferme et qu il a au moins autant de points que son adversaire.',
-          `La partie est limitee a ${cricketRounds} tours par joueur ou equipe, avec 3 fleches par tour.`,
-        ],
-      };
-    }
-
-    if (gameType === 'CAPITAL') {
-      return {
-        title: 'Regles Du Capital',
-        items: [
-          'Ordre des challenges : Capital, 20, Suite, 19, 3 a cotes, 18, 57 points, 17, Couleur, 16, Triple, 15, Double, 14, 21 ou moins, 13, Bulle ou D-Bull.',
-          'Capital : le joueur saisit directement le score total de sa visite.',
-          'Sur les challenges numeriques, seuls les segments de la cible choisie comptent. Suite, 3 a cotes et Couleur se jouent sur 3 flechettes.',
-          '57 points : il faut atteindre exactement 57, peu importe la combinaison. Si 57 est atteint avant la 3e flechette, le jeu passe directement a la suite.',
-          '21 ou moins : la visite est reussie si le total des 3 flechettes est inferieur ou egal a 21. Bulle ou D-Bull : un bull simple ou double valide le challenge.',
-          'Un challenge reussi ajoute les points marques. En cas d echec, le score du joueur est divise par 2, arrondi a l entier superieur.',
-        ],
-      };
-    }
-
-    if (gameType === 'TRIATHLON') {
-      return {
-        title: 'Regles Du Triathlon',
-        items: [
-          'Le Triathlon enchaine 3 epreuves dans cet ordre : Capital, Cricket puis 501.',
-          'Capital : 1 seule manche. En doublettes, chaque joueur joue individuellement et les scores des 2 equipiers sont additionnes pour le classement de l equipe.',
-          'Cricket : 1 seule manche. En doublettes, le Cricket se joue en equipe, comme le 501.',
-          '501 : finish au double, 1 seule manche, en individuel ou en doublettes selon le format choisi.',
-          'Le depart se decide avec une fleche a la bulle : le joueur ou l equipe le plus pres du centre commence le Triathlon.',
-          'Le score final se joue sur 100 points : chaque epreuve donne des points de resultat, puis des bonus de performance viennent completer le total.',
-          'Les bonus prennent en compte la qualite de jeu : checkout, moyenne et flechettes sur le 501 ; MPR, score et fermetures sur le Cricket ; score, regularite et penalites sur le Capital.',
-          'En cas d egalite au score final, un 501 supplementaire departage les joueurs ou les equipes. Le gagnant du tir a la bulle initial commence ce tie-break.',
-          'Les statistiques finales affichent le detail des points de resultat, des bonus et le score final du Triathlon.',
-        ],
-      };
-    }
-
-    return {
-      title: 'Regles Du Mode',
-      items: [
-        'Configure les joueurs et les options du mode avant de lancer la partie.',
-        'Les regles detaillees dependront du type de jeu selectionne.',
-      ],
-    };
-  };
-
   return (
     <div className="min-h-screen bg-[#06080d] text-white">
       <div className="absolute inset-0 bg-[radial-gradient(circle_at_top_left,rgba(249,115,22,0.18),transparent_28%),radial-gradient(circle_at_top_right,rgba(59,130,246,0.12),transparent_22%),radial-gradient(circle_at_bottom,rgba(255,255,255,0.04),transparent_35%)]" />
@@ -593,7 +217,7 @@ export const SetupView: React.FC<SetupViewProps> = ({
 
             <div className="space-y-3">
               <h2 className="text-3xl font-black uppercase tracking-[-0.04em] text-white sm:text-4xl lg:text-5xl">
-                {getTitle()}
+                {getSetupTitle(gameType)}
               </h2>
             </div>
           </div>
@@ -608,7 +232,7 @@ export const SetupView: React.FC<SetupViewProps> = ({
                 <div className="flex items-center justify-between gap-3">
                   <div>
                     <p className="text-[10px] font-black uppercase tracking-[0.24em] text-gray-500">Type De Jeu</p>
-                    <p className="mt-2 text-2xl font-black text-orange-400">{getGameName()}</p>
+                    <p className="mt-2 text-2xl font-black text-orange-400">{getGameName(gameType)}</p>
                   </div>
                   <span className="text-[10px] font-black uppercase tracking-[0.22em] text-gray-400">
                     Voir Les Regles
@@ -927,7 +551,7 @@ export const SetupView: React.FC<SetupViewProps> = ({
                   <div className="space-y-2 text-sm text-gray-300">
                     <div className="flex items-center justify-between">
                       <span>Jeu</span>
-                      <span className="font-black text-white">{getGameName()}</span>
+                      <span className="font-black text-white">{getGameName(gameType)}</span>
                     </div>
                     {gameType === 'TRIATHLON' && (
                       <>
@@ -1028,7 +652,7 @@ export const SetupView: React.FC<SetupViewProps> = ({
               <div>
                 <div className="text-[10px] font-black uppercase tracking-[0.24em] text-gray-500">Regles</div>
                 <h3 className="mt-2 text-2xl font-black uppercase tracking-[-0.04em] text-white">
-                  {getRulesContent().title}
+                  {getRulesContent(gameType, cricketRounds, checkIn, checkOut).title}
                 </h3>
               </div>
               <button
@@ -1042,7 +666,7 @@ export const SetupView: React.FC<SetupViewProps> = ({
 
             <div className="overflow-y-auto px-6 py-5">
               <div className="space-y-3">
-              {getRulesContent().items.map((item) => (
+              {getRulesContent(gameType, cricketRounds, checkIn, checkOut).items.map((item) => (
                 <div key={item} className="rounded-2xl border border-white/8 bg-[#0a1018] px-4 py-4 text-sm leading-7 text-gray-300">
                   {item}
                 </div>
@@ -1157,38 +781,6 @@ export const SetupView: React.FC<SetupViewProps> = ({
           </div>
         </div>
       )}
-    </div>
-  );
-};
-
-interface PlayerNameFieldProps {
-  label: string;
-  value: string;
-  placeholder: string;
-  onChange: (value: string) => void;
-  compact?: boolean;
-}
-
-const PlayerNameField: React.FC<PlayerNameFieldProps> = ({
-  label,
-  value,
-  placeholder,
-  onChange,
-  compact = false,
-}) => {
-  return (
-    <div className="rounded-2xl border border-white/10 bg-black/20 px-4 py-3 transition-all focus-within:border-orange-400/50">
-      <div className="mb-2 text-[10px] font-black uppercase tracking-[0.24em] text-gray-500">{label}</div>
-      <div className="rounded-xl border border-white/10 bg-white/[0.04] px-3 py-2">
-        <input
-          type="text"
-          value={value}
-          onChange={(e) => onChange(e.target.value)}
-          onFocus={(e) => e.target.select()}
-          className={`w-full bg-transparent font-black text-white placeholder:text-gray-600 focus:outline-none ${compact ? 'text-sm' : 'text-lg'}`}
-          placeholder={placeholder}
-        />
-      </div>
     </div>
   );
 };
